refactor(CLI): Rename module to unify convention
