+++ lib/cli/render-version.js
